feat: control module flow panel size
